fix: Move UserAccess enum to ui.tsx
